# pkg/server/swagger_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vigolium API",
    "description": "Vigolium — High-fidelity vulnerability scanner fusing agentic AI with native speed, modularity, and precision",
    "version": "1.0.0",
    "contact": {
      "name": "Vigolium",
      "url": "https://vigolium.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9002",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/server-info": {
      "get": {
        "tags": ["Info"],
        "summary": "Server information",
        "description": "Returns detailed server information including uptime, database driver, queue depth, and record/finding counts. Requires authentication.",
        "operationId": "getServerInfo",
        "responses": {
          "200": {
            "description": "Server information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServerInfoResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "description": "Authenticate with username and access code. Returns a bearer token for subsequent API calls.",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scans/run": {
      "post": {
        "tags": ["Scans"],
        "summary": "Run a target-based scan",
        "description": "Triggers an asynchronous target-based scan. Accepts one or more targets (domains/URLs) and optional configuration for modules, concurrency, strategy, and scope. The scan runs in the background; use GET /api/scan/status or GET /api/scans/{uuid} to monitor progress. Returns 409 if a scan is already running for the project, or 429 if the scan queue is full.",
        "operationId": "runScan",
        "parameters": [
          {
            "name": "X-Project-UUID",
            "in": "header",
            "description": "Project UUID to scope the scan",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RunScanRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Scan queued successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Scan already running for this project",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "429": {
            "description": "Scan queue is full",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scan-records": {
      "post": {
        "tags": ["Scans"],
        "summary": "Scan specific HTTP records",
        "description": "Triggers a scan on specific HTTP records identified by their UUIDs. Optionally restrict to specific modules.",
        "operationId": "scanRecords",
        "parameters": [
          {
            "name": "X-Project-UUID",
            "in": "header",
            "description": "Project UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScanRecordsRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Scan queued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scan-all-records": {
      "post": {
        "tags": ["Scans"],
        "summary": "Scan all matching HTTP records",
        "description": "Triggers a scan on all HTTP records matching the filter criteria. Supports filtering by hostname, methods, path, status codes, source, and other parameters.",
        "operationId": "scanAllRecords",
        "parameters": [
          {
            "name": "X-Project-UUID",
            "in": "header",
            "description": "Project UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScanAllRecordsRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Scan queued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scan-url": {
      "post": {
        "tags": ["Scans"],
        "summary": "Scan a single URL inline",
        "description": "Performs a synchronous scan of a single URL and returns findings inline. Useful for quick ad-hoc checks. The scan blocks until complete and returns results directly in the response.",
        "operationId": "scanURL",
        "parameters": [
          {
            "name": "X-Project-UUID",
            "in": "header",
            "description": "Project UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScanURLRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scan results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Inline findings result"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scan-request": {
      "post": {
        "tags": ["Scans"],
        "summary": "Scan a raw HTTP request inline",
        "description": "Performs a synchronous scan of a raw HTTP request and returns findings inline. Accepts a full raw HTTP request string and optional target URL override.",
        "operationId": "scanRequest",
        "parameters": [
          {
            "name": "X-Project-UUID",
            "in": "header",
            "description": "Project UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScanRequestRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scan results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Inline findings result"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scans/{uuid}/stop": {
      "post": {
        "tags": ["Scans"],
        "summary": "Stop a scan",
        "description": "Stops a running scan identified by its UUID.",
        "operationId": "stopScan",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "description": "Scan UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan stopped",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanStatusResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scans/{uuid}/pause": {
      "post": {
        "tags": ["Scans"],
        "summary": "Pause a scan",
        "description": "Pauses a running scan identified by its UUID.",
        "operationId": "pauseScan",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "description": "Scan UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan paused",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanStatusResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scans/{uuid}/resume": {
      "post": {
        "tags": ["Scans"],
        "summary": "Resume a paused scan",
        "description": "Resumes a previously paused scan identified by its UUID.",
        "operationId": "resumeScan",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "description": "Scan UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan resumed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanStatusResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scan": {
      "delete": {
        "tags": ["Scans"],
        "summary": "Cancel current scan",
        "description": "Cancels the currently running scan for the project. Requires admin role.",
        "operationId": "cancelCurrentScan",
        "parameters": [
          {
            "name": "X-Project-UUID",
            "in": "header",
            "description": "Project UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan cancelled",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanStatusResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scans/{uuid}": {
      "get": {
        "tags": ["Scans"],
        "summary": "Get scan detail",
        "description": "Returns detailed information about a specific scan including its configuration, status, statistics, and timing.",
        "operationId": "getScan",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "description": "Scan UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Full Scan database object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Scan not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Scans"],
        "summary": "Delete a scan",
        "description": "Deletes a scan record. Requires admin role.",
        "operationId": "deleteScan",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "description": "Scan UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "scan deleted" },
                    "uuid": { "type": "string", "format": "uuid" },
                    "project_uuid": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Scan not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scans": {
      "get": {
        "tags": ["Scans"],
        "summary": "List scans",
        "description": "Returns a paginated list of scans for the current project.",
        "operationId": "listScans",
        "parameters": [
          {
            "name": "X-Project-UUID",
            "in": "header",
            "description": "Project UUID",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return",
            "schema": { "type": "integer", "default": 50, "maximum": 500 }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated scan list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scan/status": {
      "get": {
        "tags": ["Scans"],
        "summary": "Get current scan status",
        "description": "Returns the current scan state for the project. Useful for polling scan progress.",
        "operationId": "getScanStatus",
        "parameters": [
          {
            "name": "project",
            "in": "query",
            "description": "Project UUID override",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "X-Project-UUID",
            "in": "header",
            "description": "Project UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Current scan status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScanStatusResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scans/queue": {
      "get": {
        "tags": ["Scans"],
        "summary": "Get scan queue info",
        "description": "Returns information about the scan queue including pending and running scans.",
        "operationId": "getScanQueue",
        "responses": {
          "200": {
            "description": "Queue information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Scan queue state"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scans/{uuid}/logs": {
      "get": {
        "tags": ["Scans"],
        "summary": "Get scan logs",
        "description": "Returns paginated scan logs for a specific scan. Supports filtering by log level and scan phase.",
        "operationId": "getScanLogs",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "description": "Scan UUID",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of log entries",
            "schema": { "type": "integer", "default": 100 }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of entries to skip",
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "level",
            "in": "query",
            "description": "Filter by log level (e.g. info, warn, error)",
            "schema": { "type": "string" }
          },
          {
            "name": "phase",
            "in": "query",
            "description": "Filter by scan phase",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project_uuid": { "type": "string", "format": "uuid" },
                    "logs": {
                      "type": "array",
                      "items": { "type": "object" }
                    },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Scan not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token obtained from POST /api/auth/login or configured API key"
      }
    },
    "schemas": {
...
      "ServerInfoResponse": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "vigolium" },
          "version": { "type": "string", "example": "1.0.0" },
          "author": { "type": "string" },
          "docs": { "type": "string" },
          "build_time": { "type": "string" },
          "commit": { "type": "string" },
          "uptime": { "type": "string", "example": "2h30m15s" },
          "service_addr": { "type": "string", "example": "0.0.0.0:9002" },
          "proxy_addr": { "type": "string", "example": "0.0.0.0:9003" },
          "queue_depth": { "type": "integer", "example": 0 },
          "total_records": { "type": "integer", "example": 1500 },
          "total_findings": { "type": "integer", "example": 42 },
          "license": { "type": "string", "example": "enterprise" },
          "demo_only": { "type": "boolean", "example": false },
          "view_only": { "type": "boolean", "example": false }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "access_code"],
        "properties": {
          "username": { "type": "string", "example": "admin" },
          "access_code": { "type": "string", "example": "my-secret-key" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "invalid request" },
          "code": { "type": "integer", "example": 400 },
          "details": { "type": "string" }
        },
        "required": ["error"]
      },
...
    }
  }
}